Fixed overflow in get_index.

+++ conways_game_of_life_impl/src/lib.rs
@@ -317,7 +317,7 @@ impl Grid1dVec {
     }
 
     fn get_index(&self, point: GridPoint) -> usize {
-        (point.0 + point.1 * self.size.0) as usize
+        point.0 as usize + point.1 as usize * self.size.0 as usize
     }
 }
